pyproject.toml
[project]
name = "learnagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
